FEAT: Change page layout by tabs
