+++ src/components/layouts/ScheduleView.tsx
@@ -2,23 +2,20 @@ import * as React from 'react'
 import Alert from '@mui/material/Alert'
 import Box from '@mui/material/Box'
 import Container from '@mui/material/Container'
-import Drawer from '@mui/material/Drawer'
 import Fab from '@mui/material/Fab'
 import Snackbar from '@mui/material/Snackbar'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faMapLocationDot } from '@fortawesome/free-solid-svg-icons'
 
-import PlanningLayout from 'components/layouts/PlaningLayout'
 import Scheduler from 'components/modules/Scheduler'
 import SchedulerHeader from 'components/modules/SchedulerHeader'
 import { useRouter } from 'hooks/useRouter'
 import { useTravelPlan } from 'hooks/useTravelPlan'
 
 type Props = {
-  open: boolean
   onClose: () => void
 }
-const ScheduleViewer: React.FC<Props> = ({ open, onClose }) => {
+const ScheduleView: React.FC<Props> = ({ onClose }) => {
   const router = useRouter()
   const [plan, planApi] = useTravelPlan()
 
@@ -44,43 +41,41 @@ const ScheduleViewer: React.FC<Props> = ({ open, onClose }) => {
   }
 
   return (
-    <Drawer open={open} anchor="bottom" onClose={onClose}>
-      <PlanningLayout>
-        <Container
-          maxWidth="md"
+    <>
+      <Container
+        maxWidth="md"
+        sx={{
+          display: 'flex',
+          flexFlow: 'column',
+          height: '100%',
+        }}>
+        <SchedulerHeader
+          plan={plan}
+          addHotel={handleAddHotel}
+          updateTitle={handleUpdate}
+        />
+        <Box sx={{ height: '100%', zIndex: 0 }}>
+          <Scheduler plan={plan} planApi={planApi} />
+        </Box>
+        <Fab
+          onClick={onClose}
+          color="primary"
           sx={{
-            display: 'flex',
-            flexFlow: 'column',
-            height: '100%',
+            position: 'fixed',
+            right: 16,
+            bottom: 16,
           }}>
-          <SchedulerHeader
-            plan={plan}
-            addHotel={handleAddHotel}
-            updateTitle={handleUpdate}
-          />
-          <Box sx={{ height: '100%', zIndex: 0 }}>
-            <Scheduler plan={plan} planApi={planApi} />
-          </Box>
-          <Fab
-            onClick={onClose}
-            color="primary"
-            sx={{
-              position: 'fixed',
-              right: 16,
-              bottom: 16,
-            }}>
-            <FontAwesomeIcon icon={faMapLocationDot} size="lg" />
-          </Fab>
-        </Container>
-      </PlanningLayout>
+          <FontAwesomeIcon icon={faMapLocationDot} size="lg" />
+        </Fab>
+      </Container>
       <Snackbar
         anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
         open={plan.waypoints?.length === 0}
         autoHideDuration={6000}>
         <Alert severity={'info'}>地図上で行きたい場所を選んでください。</Alert>
       </Snackbar>
-    </Drawer>
+    </>
   )
 }
 
-export default ScheduleViewer
+export default ScheduleView
