+++ src/components/modules/MapToolbar.tsx
@@ -1,37 +1,13 @@
 import * as React from 'react'
-import { styled } from '@mui/system'
-import SvgIcon from '@mui/material/SvgIcon'
-import Tabs from '@mui/material/Tabs'
-import Tab from '@mui/material/Tab'
-import Typography from '@mui/material/Typography'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import {
-  faCalendarWeek,
-  faEllipsis,
-  faMapLocationDot,
-  faSuitcase,
-} from '@fortawesome/free-solid-svg-icons'
 
 import SpotsCandidates from 'components/modules/SpotsCandidates'
-import ScheduleViewer from 'components/layouts/ScheduleViewer'
 import { useTravelPlan } from 'hooks/useTravelPlan'
-import PlanMenu from './PlanMenu'
 
 type Drawers = 'spots' | 'schedule'
-
-const MyTab = styled(Tab)`
-  padding: 0;
-`
 
 const MapToolbar = () => {
   const [plan] = useTravelPlan()
   const [open, setOpen] = React.useState<Drawers | null>(null)
-  const [menuAnchor, setMenuAnchor] = React.useState<null | HTMLElement>(null)
-  const [value, setValue] = React.useState(0)
-
-  const handleChange = (event: React.SyntheticEvent, newValue: number) => {
-    setValue(newValue)
-  }
 
   const handleOpen = (mode: Drawers) => () => {
     setOpen(mode)
@@ -42,44 +18,6 @@ const MapToolbar = () => {
 
   return (
     <>
-      <Tabs
-        value={value}
-        variant="fullWidth"
-        onChange={handleChange}
-        sx={{ p: 0 }}>
-        <MyTab
-          icon={
-            <SvgIcon>
-              <FontAwesomeIcon icon={faSuitcase} />
-            </SvgIcon>
-          }
-          label={<Typography variant="caption">プラン情報</Typography>}
-        />
-        <MyTab
-          icon={
-            <SvgIcon>
-              <FontAwesomeIcon icon={faMapLocationDot} />
-            </SvgIcon>
-          }
-          label={<Typography variant="caption">マップ</Typography>}
-        />
-        <MyTab
-          icon={
-            <SvgIcon>
-              <FontAwesomeIcon icon={faCalendarWeek} />
-            </SvgIcon>
-          }
-          label={<Typography variant="caption">スケジュール</Typography>}
-        />
-        <MyTab
-          icon={
-            <SvgIcon>
-              <FontAwesomeIcon icon={faEllipsis} />
-            </SvgIcon>
-          }
-          label={<Typography variant="caption">設定</Typography>}
-        />
-      </Tabs>
       {/* <Stack
         direction="row"
         justifyContent="space-around"
@@ -107,14 +45,6 @@ const MapToolbar = () => {
         spots={plan?.waypoints || []}
         onOpen={() => handleOpen('spots')}
         onClose={handleClose}
-      />
-      <ScheduleViewer open={open === 'schedule'} onClose={handleClose} />
-      <PlanMenu
-        anchorEl={menuAnchor}
-        anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
-        transformOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-        open={Boolean(menuAnchor)}
-        onClose={() => setMenuAnchor(null)}
       />
     </>
   )
