+++ src/pages/[userId]/[planId].tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import PlanView from 'components/layouts/PlanView'
+import PlanView from 'components/layouts/PlanLayout'
 import { useTravelPlan } from 'hooks/useTravelPlan'
 import { usePlans } from 'hooks/usePlan'
 import { useRouter } from 'hooks/useRouter'
