+++ src/pages/plan.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import PlanView from 'components/layouts/PlanView'
+import PlanView from 'components/layouts/PlanLayout'
 
 const FeaturedPlaces = () => {
   return <PlanView />
